Develop

+++ src/main/java/com/adventour/web/service/impl/LocationServiceImpl.java
@@ -1,51 +1,36 @@
 package com.adventour.web.service.impl;
 
 import com.adventour.web.dto.LocationDto;
+import com.adventour.web.mapper.Mapper;
 import com.adventour.web.models.Location;
 import com.adventour.web.repository.LocationRepository;
+import com.adventour.web.repository.ScheduleRepository;
 import com.adventour.web.service.LocationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class LocationServiceImpl implements  LocationService{
     private final LocationRepository locationRepository;
+    private final ScheduleRepository scheduleRepository;
+    private final Mapper mapper;
 
     @Autowired
-    public LocationServiceImpl (LocationRepository locationRepository){
+    public LocationServiceImpl (LocationRepository locationRepository, ScheduleRepository scheduleRepository, Mapper mapper){
         this.locationRepository = locationRepository;
-    }
-
-    private LocationDto mapToLocationDto(Location location) {
-        return LocationDto.builder()
-                .id(location.getId())
-                .nameLocation(location.getNameLocation())
-                .address(location.getAddress())
-                .description(location.getDescription())
-                .images(location.getImages())
-                .build();
-    }
-
-    private Location mapToLocation(LocationDto location) {
-        return Location.builder()
-                .id(location.getId())
-                .nameLocation(location.getNameLocation())
-                .address(location.getAddress())
-                .description(location.getDescription())
-                .images(location.getImages())
-                .build();
+        this.scheduleRepository = scheduleRepository;
+        this.mapper = mapper;
     }
 
     @Override
     public List<LocationDto> findAllLocation() {
 
         List<Location> locations = locationRepository.findAll();
-        return  locations.stream().map(location -> mapToLocationDto(location)).collect(Collectors.toList());
+        return  locations.stream().map(location -> mapper.mapToLocationDto(location)).collect(Collectors.toList());
     }
 
     @Override
@@ -55,7 +40,7 @@ public class LocationServiceImpl implements  LocationService{
          List<Location> locations = locationRepository.findAll();
          for(Location location : locations){
              if(location.getNameLocation().contains(search)){
-                 result.add(mapToLocationDto(location));
+                 result.add(mapper.mapToLocationDto(location));
              }
          }
         return result;
@@ -65,7 +50,7 @@ public class LocationServiceImpl implements  LocationService{
     public Location addNewLocation(LocationDto locationDto) {
 
         if(validateLocation(locationDto)){
-            Location location = mapToLocation(locationDto);
+            Location location = mapper.mapToLocation(locationDto);
             return locationRepository.save(location);
         }
         return null;
@@ -75,7 +60,7 @@ public class LocationServiceImpl implements  LocationService{
     public LocationDto getLocationById(Long id) {
         Location location = locationRepository.findById(id).orElse(null);
         if(location != null){
-            return mapToLocationDto(location);
+            return mapper.mapToLocationDto(location);
         }
         return null;
     }
@@ -83,15 +68,23 @@ public class LocationServiceImpl implements  LocationService{
     @Override
     public Location editLocation(LocationDto locationDto) {
         if(validateLocation(locationDto)){
-            Location location = mapToLocation(locationDto);
+            Location location = mapper.mapToLocation(locationDto);
             return locationRepository.save(location);
         }
         return null;
     }
 
     @Override
-    public void deleteLocation(Long locationId) {
-        locationRepository.deleteById(locationId);
+    public boolean deleteLocation(Long locationId) {
+        Location location = locationRepository.findById(locationId).orElse(null);
+        if(location != null){
+            if(scheduleRepository.findByStartLocation(location).isEmpty()
+                    && scheduleRepository.findByEndLocation(location).isEmpty()){
+                locationRepository.deleteById(locationId);
+                return true;
+            }
+        }
+        return false;
     }
 
     public boolean validateLocation(LocationDto locationDto){
